Rename .js files for consistency
