+++ scripts/bracket_orders.js
@@ -2,8 +2,8 @@ const { fetchTokenInfoFromExchange, getExchange, getSafe, buildOrders } = requir
   web3,
   artifacts
 )
-const { isPriceReasonable, areBoundsReasonable } = require("./utils/price-utils.js")(web3, artifacts)
-const { proceedAnyways } = require("./utils/user-interface-helpers")
+const { isPriceReasonable, areBoundsReasonable } = require("./utils/price_utils.js")(web3, artifacts)
+const { proceedAnyways } = require("./utils/user_interface_helpers")
 const { signAndSend, promptUser } = require("./utils/sign_and_send")(web3, artifacts)
 
 const argv = require("./utils/default_yargs")
