+++ scripts/complete_liquidity_provision.js
@@ -1,5 +1,5 @@
 const { deployFleetOfSafes } = require("./utils/trading_strategy_helpers")(web3, artifacts)
-const { isPriceReasonable, areBoundsReasonable } = require("./utils/price-utils")(web3, artifacts)
+const { isPriceReasonable, areBoundsReasonable } = require("./utils/price_utils")(web3, artifacts)
 const { sleep } = require("./utils/js_helpers")
 
 const Contract = require("@truffle/contract")
@@ -10,7 +10,7 @@ const {
   checkSufficiencyOfBalance,
 } = require("./utils/trading_strategy_helpers")(web3, artifacts)
 const { signAndSend, promptUser } = require("./utils/sign_and_send")(web3, artifacts)
-const { proceedAnyways } = require("./utils/user-interface-helpers")(web3, artifacts)
+const { proceedAnyways } = require("./utils/user_interface_helpers")(web3, artifacts)
 
 const { toErc20Units } = require("./utils/printing_tools")
 const assert = require("assert")
