+++ scripts/utils/user_interface_helpers.js
