+++ test/printing_tools.js
